fix: babysit round 3, 7 findings (7 coderabbit, 0 ci)
Narrow the pytest PLR0917 exemption to the seven modules whose signatures
pytest or Hypothesis fills by name, matching the enumerated controller
treatment; the two TYPE_CHECKING conformance stubs take a bare '*'.

Split STEP_EXHAUSTED from STEP_FAILED so a step abandoned on its turn
budget no longer reaches the quality pipeline wearing the same reason as
one the provider botched, which now classifies INCORRECT rather than
neutral.

Promote the mini-ReAct turn pipeline into PlanStepTurnMixin and move
call_planner into plan_helpers, so both loops share one turn body and one
planner body instead of byte-identical copies. run_progress_summary now
runs through run_planner_turn too.

Resync the plan history on the three paths that finalise straight from
all_plans, so a terminal result no longer loses a FAILED or COMPLETED
step status.
