+++ data/architecture_report.json
@@ -1664,7 +1664,7 @@
     "synthorg.engine.workflow.definition": 42,
     "synthorg.engine.workflow.enums": 40,
     "synthorg.engine.workflow.sprint_lifecycle": 20,
-    "synthorg.execution.turn": 32,
+    "synthorg.execution.turn": 31,
     "synthorg.hr.enums": 30,
     "synthorg.hr.performance.models": 47,
     "synthorg.hr.performance.tracker": 29,
@@ -1705,7 +1705,7 @@
     "synthorg.meta.toolsmith.models": 21,
     "synthorg.notifications.dispatcher": 23,
     "synthorg.notifications.models": 24,
-    "synthorg.observability": 1938,
+    "synthorg.observability": 1939,
     "synthorg.observability.background_tasks": 30,
     "synthorg.observability.events.api": 235,
     "synthorg.observability.events.approval_gate": 27,
@@ -1714,7 +1714,7 @@
     "synthorg.observability.events.communication": 26,
     "synthorg.observability.events.config": 22,
     "synthorg.observability.events.evolution": 24,
-    "synthorg.observability.events.execution": 48,
+    "synthorg.observability.events.execution": 49,
     "synthorg.observability.events.hr": 39,
     "synthorg.observability.events.integrations": 94,
     "synthorg.observability.events.mcp": 67,
@@ -1827,8 +1827,8 @@
       "loc": 764
     },
     "synthorg.engine.hybrid_loop:HybridLoop": {
-      "lcom4": 1,
-      "loc": 745
+      "lcom4": 3,
+      "loc": 591
     },
     "synthorg.engine.initiative.rollup:ProjectRollupService": {
       "lcom4": 1,
