+++ data/runtime_stats.yaml
@@ -1,9 +1,9 @@
 schema_version: 1
-last_generated_utc: '2026-07-26T21:38:58Z'
-generator_revision: 4b446993a
+last_generated_utc: '2026-07-26T22:18:39Z'
+generator_revision: ad6789250
 stats:
   tests:
-    raw: 42630
+    raw: 42635
     rounded: 42000
     display: 42,000+
   providers_curated:
