+++ pyproject.toml
@@ -377,10 +377,6 @@ max-complexity = 8
     "SLF001",  # private member access in tests
     "PLC0415", # local imports in test functions
     "TC",      # type-checking imports (fixture type hints)
-    # pytest injects fixtures by parameter name, so a test's positional
-    # count is chosen by the fixture graph rather than by any caller. The
-    # rule has nothing to act on: there is no call site to convert.
-    "PLR0917", # too-many-positional-arguments (fixture injection)
     # Tests carry parametrize values, fixtures, and scaffolding that the
     # tightened ruff rules don't fit; relaxations below.
     "BLE001",  # blind-except in tests (acceptable in test infra)
@@ -471,6 +467,19 @@ max-complexity = 8
 "src/synthorg/api/controllers/project_docs.py" = ["PLR0917"]
 "src/synthorg/api/controllers/project_knowledge.py" = ["PLR0917"]
 "src/synthorg/api/controllers/tasks.py" = ["PLR0917"]
+# Test modules whose signatures pytest or Hypothesis fills by parameter
+# name: a fixture graph or a ``@given`` / ``@pytest.mark.parametrize``
+# decorator chooses the positional count, so there is no call site to
+# convert. Enumerated per module rather than globbed over tests/**, for
+# the same reason as the controllers above: a blanket test-tree ignore
+# would let an ordinary helper skip the cap silently.
+"tests/unit/api/conftest.py" = ["PLR0917"]
+"tests/unit/api/controllers/test_agent_identity_versions.py" = ["PLR0917"]
+"tests/unit/core/test_task_properties.py" = ["PLR0917"]
+"tests/unit/memory/procedural/capture/test_hybrid_capture.py" = ["PLR0917"]
+"tests/unit/meta/mcp/test_handlers_infrastructure.py" = ["PLR0917"]
+"tests/unit/settings/test_resolver_bridge_configs.py" = ["PLR0917"]
+"tests/unit/settings/test_service.py" = ["PLR0917"]
 "src/synthorg/tools/browser/**/*.py" = [
     # The BrowserTool catches its own BrowserDomainError chain at the
     # execute() boundary and emits a structured ToolExecutionResult.
