+++ src/synthorg/engine/hybrid_loop.py
@@ -32,7 +32,6 @@ from synthorg.observability.events.execution import (
     EXECUTION_LOOP_ERROR,
     EXECUTION_LOOP_START,
     EXECUTION_LOOP_TERMINATED,
-    EXECUTION_LOOP_TURN_COMPLETE,
     EXECUTION_PLAN_CREATED,
     EXECUTION_PLAN_STEP_COMPLETE,
     EXECUTION_PLAN_STEP_START,
@@ -41,8 +40,6 @@ from synthorg.providers.enums import MessageRole
 from synthorg.providers.models import (
     ChatMessage,
     CompletionConfig,
-    CompletionResponse,
-    ToolDefinition,
 )
 from synthorg.providers.protocol import CompletionProvider
 from synthorg.tools.protocol import ToolInvokerProtocol
@@ -50,13 +47,10 @@ from synthorg.tools.protocol import ToolInvokerProtocol
 from .hybrid.replan_helpers import attempt_replan, do_replan, run_progress_summary
 from .hybrid.step_helpers import (
     build_step_message,
-    call_planner,
-    handle_step_completion,
     truncate_plan,
     warn_insufficient_budget,
 )
 from .hybrid_models import HybridLoopConfig
-from .intervention.loop_hook import check_steering
 from .loop_cancellation import check_task_cancelled
 from .loop_control_helpers import (
     check_budget,
@@ -66,13 +60,9 @@ from .loop_control_helpers import (
 )
 from .loop_helpers import (
     build_result,
-    check_response_errors,
     classify_step,
-    classify_turn,
     get_tool_definitions,
-    make_turn_record,
     notify_turn_observer,
-    response_to_message,
 )
 from .loop_protocol import (
     BudgetChecker,
@@ -82,18 +72,9 @@ from .loop_protocol import (
     TerminationReason,
     TurnObserver,
 )
-from .loop_streaming import (
-    _TurnInterrupted,
-    fold_interrupt_usage,
-    run_provider_turn,
-)
-from .loop_tool_execution import (
-    clear_last_turn_tool_calls,
-    execute_tool_calls,
-)
 from .plan_helpers import (
+    call_planner,
     clear_superseded_directive,
-    invoke_checkpoint_callback,
     update_step_status,
 )
 from .plan_loop_context import (
@@ -109,11 +90,12 @@ from .plan_models import (
     StepStatus,
 )
 from .plan_parsing import _PLANNING_PROMPT
+from .plan_step_turn import PlanStepTurnMixin
 
 logger = get_logger(__name__)
 
 
-class HybridLoop:
+class HybridLoop(PlanStepTurnMixin):
     """Hybrid Plan + ReAct execution loop.
 
     Plans, then executes each step as a mini-ReAct loop with a
@@ -353,9 +335,7 @@ class HybridLoop:
                 signals,
                 state,
                 step_start,
-                TerminationReason.COMPLETED
-                if step_result.step_succeeded
-                else TerminationReason.MAX_TURNS,
+                step_result.signal_reason,
             )
 
             if step_result.step_succeeded:
@@ -760,16 +740,18 @@ class HybridLoop:
             if isinstance(stag_outcome, tuple):
                 state.ctx, step_corrections = stag_outcome
 
-        # Loop exited without step completion
+        # Loop exited without step completion: the run's turn budget or the
+        # per-step one ran out, which is a different failure from a step the
+        # model concluded unsuccessfully.
         if not state.ctx.has_turns_remaining:
-            return StepTurnOutcome.STEP_FAILED
+            return StepTurnOutcome.STEP_EXHAUSTED
         logger.warning(
             EXECUTION_HYBRID_STEP_TURN_LIMIT,
             execution_id=state.ctx.execution_id,
             step_number=step.step_number,
             max_turns_per_step=self._config.max_turns_per_step,
         )
-        return StepTurnOutcome.STEP_FAILED
+        return StepTurnOutcome.STEP_EXHAUSTED
 
     async def _compact(self, ctx: AgentContext) -> AgentContext:
         """Run context compaction at turn boundaries.
@@ -785,179 +767,6 @@ class HybridLoop:
             ctx.turn_count,
         )
         return compacted if compacted is not None else ctx
-
-    async def _run_step_turn(
-        self,
-        run: StepRunContext,
-        state: StepRunState,
-        tool_defs: list[ToolDefinition] | None,
-    ) -> StepTurnOutcome | ExecutionResult:
-        """Execute a single turn within a step's mini-ReAct sub-loop.
-
-        Returns:
-            :attr:`StepTurnOutcome.CONTINUE` to keep the sub-loop running
-            (also the re-issue path after a mid-turn steering REDIRECT),
-            ``STEP_SUCCEEDED`` / ``STEP_FAILED`` once the step concludes, or
-            an :class:`ExecutionResult` to terminate the run: shutdown,
-            budget, cancellation, a provider error, or an approval-gate or
-            tool-failure outcome from the tool-call arm.
-        """
-        blocked = await self._pre_turn_checks(run, state)
-        if blocked is not None:
-            return blocked
-
-        turn_number = state.ctx.turn_count + 1
-        outcome = await run_provider_turn(
-            state.ctx,
-            run.provider,
-            run.executor_model,
-            tool_defs=tool_defs,
-            config=run.completion_config,
-            turn_number=turn_number,
-            turns=state.turns,
-            streaming_enabled=run.streaming_enabled,
-            cancellation_checker=run.task_cancellation_checker,
-            steering_inbox=self._steering_inbox,
-        )
-        if isinstance(outcome, ExecutionResult):
-            return outcome
-        if isinstance(outcome, _TurnInterrupted):
-            # Re-issue the step turn: continuing the mini-sub-loop lets its
-            # top-of-turn steering check adopt the REDIRECT the interrupt
-            # fired for.
-            state.ctx = fold_interrupt_usage(state.ctx, outcome)
-            return StepTurnOutcome.CONTINUE
-
-        return await self._finish_turn(run, state, outcome, turn_number)
-
-    async def _pre_turn_checks(
-        self,
-        run: StepRunContext,
-        state: StepRunState,
-    ) -> ExecutionResult | None:
-        """Run the pre-LLM guards and adopt any pending steering directive.
-
-        Returns:
-            A terminal :class:`ExecutionResult` when shutdown, budget, or
-            cancellation trips, or ``None`` when the turn may proceed.
-        """
-        shutdown_result = check_shutdown(state.ctx, run.shutdown_checker, state.turns)
-        if shutdown_result is not None:
-            return shutdown_result
-        budget_result = check_budget(state.ctx, run.budget_checker, state.turns)
-        if budget_result is not None:
-            return budget_result
-        cancel_result = await check_task_cancelled(
-            state.ctx, run.task_cancellation_checker, state.turns
-        )
-        if cancel_result is not None:
-            return cancel_result
-        # Adopt pending steering directives before the LLM call so the
-        # operator's constraint is in context for this step's turn.
-        steered = await check_steering(state.ctx, self._steering_inbox)
-        if steered is not None:
-            state.ctx = steered
-        return None
-
-    async def _finish_turn(
-        self,
-        run: StepRunContext,
-        state: StepRunState,
-        response: CompletionResponse,
-        turn_number: int,
-    ) -> StepTurnOutcome | ExecutionResult:
-        """Record a completed provider turn and dispatch on its tool calls.
-
-        Returns:
-            :attr:`StepTurnOutcome.CONTINUE` when tool calls ran and the
-            sub-loop should issue another turn, ``STEP_SUCCEEDED`` /
-            ``STEP_FAILED`` when this turn concluded the step, or an
-            :class:`ExecutionResult` for a provider error or a terminal
-            tool-execution outcome.
-        """
-        state.turns.append(
-            make_turn_record(
-                turn_number,
-                response,
-                call_category=classify_turn(turn_number, response, state.ctx),
-                provider_metadata=response.provider_metadata,
-            )
-        )
-
-        error = check_response_errors(
-            state.ctx,
-            response,
-            turn_number,
-            state.turns,
-        )
-        if error is not None:
-            return error
-
-        state.ctx = state.ctx.with_turn_completed(
-            response.usage,
-            response_to_message(response),
-        )
-        logger.info(
-            EXECUTION_LOOP_TURN_COMPLETE,
-            execution_id=state.ctx.execution_id,
-            turn=turn_number,
-            finish_reason=response.finish_reason.value,
-            tool_call_count=len(response.tool_calls),
-        )
-
-        await invoke_checkpoint_callback(
-            run.checkpoint_callback,
-            state.ctx,
-            turn_number,
-        )
-
-        if not response.tool_calls:
-            state.ctx, step_ok = handle_step_completion(
-                state.ctx, response, turn_number
-            )
-            return StepTurnOutcome.from_success(success=step_ok)
-
-        return await self._handle_step_tool_calls(
-            run,
-            state,
-            response,
-            turn_number,
-        )
-
-    async def _handle_step_tool_calls(
-        self,
-        run: StepRunContext,
-        state: StepRunState,
-        response: CompletionResponse,
-        turn_number: int,
-    ) -> StepTurnOutcome | ExecutionResult:
-        """Check shutdown and execute tool calls for a step turn.
-
-        Returns:
-            :attr:`StepTurnOutcome.CONTINUE` once the tool outputs are
-            appended to ``state.ctx``, or an :class:`ExecutionResult` when
-            shutdown fires or tool execution terminates the loop (a missing
-            invoker, a tool failure, or an approval-gate escalation).
-        """
-        shutdown_result = check_shutdown(state.ctx, run.shutdown_checker, state.turns)
-        if shutdown_result is not None:
-            clear_last_turn_tool_calls(state.turns)
-            return shutdown_result.model_copy(
-                update={"turns": tuple(state.turns)},
-            )
-
-        result = await execute_tool_calls(
-            state.ctx,
-            run.tool_invoker,
-            response,
-            turn_number,
-            state.turns,
-            approval_gate=self._approval_gate,
-        )
-        if isinstance(result, ExecutionResult):
-            return result
-        state.ctx = result
-        return StepTurnOutcome.CONTINUE
 
     # -- Utilities ---------------------------------------------------------
