+++ web/src/api/types/artifacts.ts
@@ -5,11 +5,8 @@ export type { Artifact, CreateArtifactRequest } from './dtos.gen'
 /**
  * Frontend-only query filter (not on the wire as a Pydantic DTO).
  *
- * Every field here is a query parameter `GET /artifacts` actually accepts.
- * Project narrowing is deliberately absent: the endpoint has no project
- * filter, so a `project_id` here would be dropped server-side with no
- * error. `useArtifactsData` narrows by `Artifact.project_id` on the client
- * instead, using the field the response DTO does carry.
+ * No project field: the endpoint has no project filter, so declaring one
+ * here would let a caller pass a narrowing the server silently drops.
  */
 export interface ArtifactFilters {
   task_id?: string
